bottom them top ll study
When this gets rendered on the screen, LinearLayout will make two passes through its children. On the first pass, it asks for how much space each child wants, based upon the android:layout_height values (since this is a vertical LinearLayout). The first Button will ask for 0 pixels of space, while the bottom Button will ask for however much it needs for its content. Then, the second pass of the LinearLayout through its children asks for their weights. The first Button has a weight of 1, while the bottom Button has the default weight of 0. As a result, the top Button gets 1/1 = 100% of all pixels left over from what the first pass used. This causes the top Button to become tall, shoving the bottom Button to the bottom of the screen.

1. height
2. weight